Increase NPTH diameter for AD-121F2 seven-segment display.
The datasheet says the little plastic pegs have a diameter of 1.0 mm,
so I thought that a 1.2 mm hole would be big enough.  When I got my
board back, it didn't fit.  So, I increased the hole diameter to 1.5mm
and ran a test board.  It seems to fit nicely now.  Although it's a
bit unsatisfying that this was determined empirically, I think it's
the number we should go with.